+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Changing the predefined search filters ([UIEUS-450](https://folio-org.atlassian.net/browse/UIEUS-450))
 * Integration of the new term COUNTER alongside SUSHI ([UIEUS-451](https://folio-org.atlassian.net/browse/UIEUS-451))
 * Add icon for failed reports with error code 3030 ([UIEUS-440](https://folio-org.atlassian.net/browse/UIEUS-440))
+* Restructure monthpicker for fixing validation ([UIEUS-457](https://folio-org.atlassian.net/browse/UIEUS-457))
 
 ## [11.0.1](https://github.com/folio-org/ui-erm-usage/tree/v11.0.1) (2025-04-11)
 * Fix settings aggregator form ([UIEUS-411](https://folio-org.atlassian.net/browse/UIEUS-411))
