+++ src/components/HarvestingConfiguration/Fields/HarvestingEndField.js
@@ -1,13 +1,10 @@
-import { Field } from 'react-final-form';
 import { FormattedMessage } from 'react-intl';
 
 import Monthpicker from '../../../util/Monthpicker';
 
 const HarvestingEndField = () => (
-  <Field
+  <Monthpicker
     backendDateFormat="YYYY-MM"
-    component={Monthpicker}
-    // dateFormat="YYYY-MM"
     id="input-harvestingEnd"
     name="harvestingConfig.harvestingEnd"
     textLabel={<FormattedMessage id="ui-erm-usage.udpHarvestingConfig.harvestingEnd" />}
