+++ src/components/HarvestingConfiguration/Fields/HarvestingStartField.js
@@ -1,15 +1,12 @@
 import { PropTypes } from 'prop-types';
-import { Field } from 'react-final-form';
 import { FormattedMessage } from 'react-intl';
 
 import Monthpicker from '../../../util/Monthpicker';
 
 const HarvestingStartField = ({ isRequired }) => {
   return (
-    <Field
+    <Monthpicker
       backendDateFormat="YYYY-MM"
-      component={Monthpicker}
-      data={isRequired ? 1 : 0}
       id="input-harvestingStart"
       isRequired={isRequired}
       name="harvestingConfig.harvestingStart"
